fix(api): refactor config and api client
